// src/shared/api/generated/model/platform.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Amazing EKB Hub API
 * MVP backend API (OpenAPI-first)
 * OpenAPI spec version: 1.0.0
 */

/**
 * Платформа, на которой опубликован материал.
 */
export type Platform = 'dzen' | 'telegram' | 'instagram';